+++ FirebaseAI/CHANGELOG.md
@@ -2,6 +2,9 @@
 - [feature] Added support for the Code Execution tool, which enables the model
   to generate and run code to perform complex tasks like solving mathematical
   equations or visualizing data. (#15280)
+- [fixed] Fixed a decoding error when generating images with the
+  `gemini-2.5-flash-image-preview` model using `generateContentStream` or
+  `sendMessageStream` with the Gemini Developer API. (#15262)
 
 # 12.2.0
 - [feature] Added support for returning thought summaries, which are synthesized